requirements.txt
gdown==3.3.1
jupyter==1.0.0 
jupyter_contrib_nbextensions==0.5.1
kaggle==1.5.10
numpy==1.19.2
pandas==1.1.3
pyyaml==5.3.1
scikit-learn==0.23.2
scipy==1.5.3
tqdm==4.50.2

